refactor : 모든 Player에게 같은 initialDeal이 주어지지 않도록 변경
